requirements.txt
dramatiq==1.16.0
redis==5.0.1
watchdog==4.0.1
pymupdf==1.24.9
Pillow==10.4.0
pytesseract==0.3.13
python-dotenv==1.0.1
loguru==0.7.2
